Testing basic functionality of splitting setup code
