+++ .setup/bin/sample_courses/main.py
@@ -51,61 +51,20 @@ import pdb
 import docker
 import random
 from tempfile import TemporaryDirectory
-
 from submitty_utils import dateutils
 
-from ruamel.yaml import YAML
 from sqlalchemy import create_engine, Table, MetaData, bindparam, select, join, func
 
-yaml = YAML(typ='safe')
-
-CURRENT_PATH = os.path.dirname(os.path.realpath(__file__))
-SETUP_DATA_PATH = os.path.join(CURRENT_PATH, "..", "data")
-
-# Default values, will be overwritten in `main()` if corresponding arguments are supplied
-SUBMITTY_INSTALL_DIR = "/usr/local/submitty"
-SUBMITTY_DATA_DIR = "/var/local/submitty"
-SUBMITTY_REPOSITORY = os.path.join(SUBMITTY_INSTALL_DIR, "GIT_CHECKOUT/Submitty")
-MORE_EXAMPLES_DIR = os.path.join(SUBMITTY_INSTALL_DIR, "more_autograding_examples")
-TUTORIAL_DIR = os.path.join(SUBMITTY_INSTALL_DIR, "GIT_CHECKOUT/Tutorial", "examples")
-
-DB_HOST = "localhost"
-DB_PORT = 5432
-DB_USER = "submitty_dbuser"
-DB_PASS = "submitty_dbuser"
-
-# used for constructing the url to clone repos for vcs gradeables
-# with open(os.path.join(SUBMITTY_INSTALL_DIR, "config", "submitty.json")) as submitty_config:
-#     submitty_config_json = json.load(submitty_config)
-SUBMISSION_URL = 'submitty_config_json["submission_url"]'
-VCS_FOLDER = os.path.join(SUBMITTY_DATA_DIR, 'vcs', 'git')
-
-DB_ONLY = False
-NO_SUBMISSIONS = False
-NO_GRADING = False
-
-NOW = dateutils.get_current_time()
-
+# if you need to modify any global variables, change this to import file as name
+from sample_courses import *
+from utils import *
+from models import *
 
 def main():
     """
     Main program execution. This gets us our commandline arguments, reads in the data files,
     and then sets us up to run the create methods for the users and courses.
     """
-    global DB_ONLY, NO_SUBMISSIONS, NO_GRADING
-    global DB_HOST, DB_PORT, DB_USER, DB_PASS
-    global SUBMITTY_INSTALL_DIR, SUBMITTY_DATA_DIR, SUBMITTY_REPOSITORY
-    global MORE_EXAMPLES_DIR, TUTORIAL_DIR
-
-    args = parse_args()
-    DB_ONLY = args.db_only
-    NO_SUBMISSIONS = args.no_submissions
-    NO_GRADING = args.no_grading
-    SUBMITTY_INSTALL_DIR = args.install_dir
-    SUBMITTY_DATA_DIR = args.data_dir
-    SUBMITTY_REPOSITORY = os.path.join(SUBMITTY_INSTALL_DIR, "GIT_CHECKOUT/Submitty")
-    MORE_EXAMPLES_DIR = os.path.join(SUBMITTY_INSTALL_DIR, "more_autograding_examples")
-    TUTORIAL_DIR = os.path.join(SUBMITTY_INSTALL_DIR, "GIT_CHECKOUT/Tutorial", "examples")
 
     if not os.path.isdir(SUBMITTY_INSTALL_DIR):
         raise SystemError(f"The following directory does not exist: {SUBMITTY_INSTALL_DIR}")
@@ -284,511 +243,6 @@ def main():
     if not NO_GRADING:
         # queue up all of the newly created submissions to grade!
         os.system(f"{SUBMITTY_INSTALL_DIR}/bin/regrade.py --no_input {SUBMITTY_DATA_DIR}/courses/")
-
-
-def get_random_text_from_file(filename):
-    line = ""
-    with open(os.path.join(SETUP_DATA_PATH, 'random', filename)) as comment:
-        line = next(comment)
-        for num, alternate_line in enumerate(comment):
-            if random.randrange(num + 2):
-                continue
-            line = alternate_line
-    return line.strip()
-
-
-def generate_random_user_id(length=15):
-    return ''.join(random.choice(string.ascii_lowercase + string.ascii_uppercase + string.digits) for _ in range(length))
-
-
-def generate_random_ta_comment():
-    return get_random_text_from_file('TAComment.txt')
-
-
-def generate_random_ta_note():
-    return get_random_text_from_file('TANote.txt')
-
-
-def generate_random_student_note():
-    return get_random_text_from_file('StudentNote.txt')
-
-def generate_pronouns():
-    pronoun_num = random.random()
-    if pronoun_num <= .05: 
-        pronoun_list = ["Ze/Zir","Xe/Xem", "Ne/Nem", "Vi/Vir", "Ne/Nir" "Nix/Nix", "Xy/Xyr", "Zhe/Zhim"]
-        return random.choice(pronoun_list)
-    elif pronoun_num <= .30:
-        return ""
-    elif pronoun_num <= .60:
-        return "She/Her"
-    elif pronoun_num <= .70:
-        return "They/Them"
-    else: return "He/Him"
-
-
-def generate_random_marks(default_value, max_value):
-    with open(os.path.join(SETUP_DATA_PATH, 'random', 'marks.yml')) as f:
-        marks_yml = yaml.load(f)
-    if default_value == max_value and default_value > 0:
-        key = 'count_down'
-    else:
-        key = 'count_up'
-    marks = []
-    mark_list = random.choice(marks_yml[key])
-    for i in range(len(mark_list)):
-        marks.append(Mark(mark_list[i], i))
-    return marks
-
-
-def generate_versions_to_submit(num=3, original_value=3):
-    if num == 1:
-        return original_value
-    if random.random() < 0.3:
-        return generate_versions_to_submit(num-1, original_value)
-    else:
-        return original_value-(num-1)
-
-
-def generate_probability_space(probability_dict, default=0):
-    """
-    This function takes in a dictionary whose key is the probability (decimal less than 1),
-    and the value is the outcome (whatever the outcome is).
-    """
-    probability_counter = 0
-    target_random = random.random()
-    prev_random_counter = 0
-    for key in sorted(probability_dict.keys()):
-        value = probability_dict[key]
-        probability_counter += key
-        if probability_counter >= target_random and target_random > prev_random_counter:
-            return value
-        prev_random_counter = probability_counter
-    return default
-
-
-def generate_random_users(total, real_users):
-    """
-
-    :param total:
-    :param real_users:
-    :return:
-    :rtype: list[User]
-    """
-    with open(os.path.join(SETUP_DATA_PATH, 'random', 'familyNames.txt')) as family_file, \
-            open(os.path.join(SETUP_DATA_PATH, 'random', 'maleGivenNames.txt')) as male_file, \
-            open(os.path.join(SETUP_DATA_PATH, 'random', 'womenGivenNames.txt')) as woman_file:
-        family_names = family_file.read().strip().split()
-        male_names = male_file.read().strip().split()
-        women_names = woman_file.read().strip().split()
-
-    users = []
-    user_ids = []
-    anon_ids = []
-    with open(os.path.join(SETUP_DATA_PATH, "random_users.txt"), "w") as random_users_file:
-        for i in range(total):
-            if random.random() < 0.5:
-                given_name = random.choice(male_names)
-            else:
-                given_name = random.choice(women_names)
-            family_name = random.choice(family_names)
-            user_id = family_name.replace("'", "")[:5] + given_name[0]
-            user_id = user_id.lower()
-            anon_id = generate_random_user_id(15)
-            # create a binary string for the numeric ID
-            numeric_id = '{0:09b}'.format(i)
-            while user_id in user_ids or user_id in real_users:
-                if user_id[-1].isdigit():
-                    user_id = user_id[:-1] + str(int(user_id[-1]) + 1)
-                else:
-                    user_id = user_id + "1"
-            if anon_id in anon_ids:
-                anon_id = generate_random_user_id()
-            new_user = User({"user_id": user_id,
-                             "user_numeric_id": numeric_id,
-                             "user_givenname": given_name,
-                             "user_familyname": family_name,
-                             "user_pronouns": generate_pronouns(),
-                             "user_group": 4,
-                             "courses": dict()})
-            new_user.create()
-            user_ids.append(user_id)
-            users.append(new_user)
-            anon_ids.append(anon_id)
-            random_users_file.write(user_id + "\n")
-    return users
-
-
-def load_data_json(file_name):
-    """
-    Loads json file from the .setup/data directory returning the parsed structure
-    :param file_name: name of file to load
-    :return: parsed JSON structure from loaded file
-    """
-    file_path = os.path.join(SETUP_DATA_PATH, file_name)
-    if not os.path.isfile(file_path):
-        raise IOError("Missing the json file .setup/data/{}".format(file_name))
-    with open(file_path) as open_file:
-        json_file = json.load(open_file)
-    return json_file
-
-
-def load_data_yaml(file_path):
-    """
-    Loads yaml file from the .setup/data directory returning the parsed structure
-    :param file_path: name of file to load
-    :return: parsed YAML structure from loaded file
-    """
-    if not os.path.isfile(file_path):
-        raise IOError("Missing the yaml file {}".format(file_path))
-    with open(file_path) as open_file:
-        yaml_file = yaml.load(open_file)
-    return yaml_file
-
-
-def user_exists(user):
-    """
-    Checks to see if the user exists on the linux file system. We can use this to delete a user
-    so that we can recreate them which avoids users having left over data from a previous run of
-    setting up the sample courses.
-    :param user: string to check if user exists
-    :return: boolean on if user exists or not
-    """
-    try:
-        pwd.getpwnam(user)
-        return True
-    except KeyError:
-        return False
-
-
-def group_exists(group):
-    """
-    Checks to see if the group exists on the linux file system so that we don't try to create
-    groups that already exist.
-
-    :param group: string to check if group exists
-    :return: boolean on if group exists or not
-    """
-    try:
-        grp.getgrnam(group)
-        return True
-    except KeyError:
-        return False
-
-
-def create_group(group):
-    """
-    Creates the group on the system, adding some base users to the group as well that are necessary
-    for the system to function and are not defined within the users.yml file.
-    :param group: name of the group to create
-    """
-    if not group_exists(group):
-        os.system("groupadd {}".format(group))
-
-    if group == "sudo":
-        return
-
-
-def add_to_group(group, user_id):
-    """
-    Adds the user to the specified group, creating the group if it does not exist.
-    :param group:
-    :param user_id:
-    """
-    create_group(group)
-    os.system("usermod -a -G {} {}".format(group, user_id))
-
-
-def get_php_db_password(password):
-    """
-    Generates a password to be used within the site for database authentication. The password_hash
-    function (http://php.net/manual/en/function.password-hash.php) generates us a nice secure
-    password and takes care of things like salting and hashing.
-    :param password:
-    :return: password hash to be inserted into the DB for a user
-    """
-    proc = subprocess.Popen(
-        ["php", "-r", "print(password_hash('{}', PASSWORD_DEFAULT));".format(password)],
-        stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    (out, err) = proc.communicate()
-    return out.decode('utf-8')
-
-
-def get_current_semester():
-    """
-    Given today's date, generates a three character code that represents the semester to use for
-    courses such that the first half of the year is considered "Spring" and the last half is
-    considered "Fall". The "Spring" semester  gets an S as the first letter while "Fall" gets an
-    F. The next two characters are the last two digits in the current year.
-    :return:
-    """
-    today = datetime.today()
-    semester = "f" + str(today.year)[-2:]
-    if today.month < 7:
-        semester = "s" + str(today.year)[-2:]
-    return semester
-
-
-def parse_args():
-    """
-    Parses out the arguments that might be passed to this script as it's run as a commandline
-    application.
-    :return: parsed args from the argparse module
-    """
-    parser = argparse.ArgumentParser(
-        description="Sets up the sample courses as well as creating the necessary users for the "
-                    "course as needed. It reads in the courses.json and users.json files from the "
-                    ".setup/data directory to determine what courses/users are allowed and then "
-                    "either adds all or just a few depending on what gets passed to this script")
-
-    parser.add_argument("--db_only", action='store_true', default=False)
-    parser.add_argument("--no_submissions", action='store_true', default=False)
-    parser.add_argument("--no_grading", action='store_true', default=False)
-    parser.add_argument("--users_path", default=os.path.join(SETUP_DATA_PATH, "users"),
-                        help="Path to folder that contains .yml files to use for user creation. Defaults to "
-                             "../data/users")
-    parser.add_argument("--submission_url", type=str, default="",help="top level url for the website")
-    parser.add_argument("--courses_path", default=os.path.join(SETUP_DATA_PATH, "courses"),
-                        help="Path to the folder that contains .yml files to use for course creation. Defaults to "
-                             "../data/courses")
-    parser.add_argument("--install_dir", type=str, default=SUBMITTY_INSTALL_DIR, help="install path of submitty")
-    parser.add_argument("--data_dir", type=str, default=SUBMITTY_DATA_DIR, help="data path of submitty")
-    parser.add_argument("course", nargs="*",
-                        help="course code to build. If no courses are passed in, then it'll use "
-                             "all courses in courses.json")
-    return parser.parse_args()
-
-
-def create_gradeable_submission(src, dst):
-    """
-    Given a source and a destination, copy the files from the source to the destination. First, before
-    copying, we check if the source is a directory, if it is, then we zip the contents of this to a temp
-    zip file (stored in /tmp) and store the path to this newly created zip as our new source.
-
-    At this point, (for all uploads), we check if our source is a zip (by just checking file extension is
-    a .zip), then we will extract the contents of the source (using Shutil) to the destination, else we
-    just do a simple copy operation of the source file to the destination location.
-
-    At this point, if we created a zip file (as part of that first step), we remove it from the /tmp directory.
-
-    :param src: path of the file or directory we want to use for this submission
-    :type src: str
-    :param dst: path to the folder where we should copy the submission to
-    :type src: str
-    """
-    zip_dst = None
-    if os.path.isdir(src):
-        zip_dst = os.path.join("/tmp", str(uuid.uuid4()))
-        zip_dst = shutil.make_archive(zip_dst, 'zip', src)
-        src = zip_dst
-
-    if src[-3:] == "zip":
-        shutil.unpack_archive(src, dst)
-    else:
-        shutil.copy(src, dst)
-
-    if zip_dst is not None and isinstance(zip_dst, str):
-        os.remove(zip_dst)
-
-def create_pdf_annotations(file_name, file_path, src, dst, grader_id):
-    """
-    Specifically designed helper function that copies a annotation from the source to the destination.
-    The source annotation need to be modified to reflect:
-        the file that the annotations belongs to
-        the grader that is responsible for the annotation
-
-    :param file_name: encoded file name
-    :type src: str
-    :param file_path: anonymous file path
-    :type src: str
-    :param src: path of the file or directory we want to use for this annotation
-    :type src: str
-    :param dst: path to the folder where we should copy the annotation to
-    :type src: str
-    :param grader_id: grader of the annotation
-    :type src: str
-    """
-    with open(src, 'r') as open_file:
-        annotation_json = json.load(open_file)
-        annotation_json['file_path'] = file_path
-        annotation_json['grader_id'] = grader_id
-        for annotation in annotation_json['annotations']:
-            annotation['userId'] = grader_id
-
-    with open(os.path.join(dst, file_name), 'w') as f:
-        json.dump(annotation_json, f, indent = 2)
-
-def commit_submission_to_repo(user_id, src_file, repo_path, vcs_subdirectory):
-    # a function to commit and push a file to a user's submitty-hosted repository
-    my_cwd = os.getcwd()
-    with TemporaryDirectory() as temp_dir:
-        os.chdir(temp_dir)
-        os.system(f'git clone {SUBMITTY_DATA_DIR}/vcs/git/{repo_path}')
-        os.chdir(os.path.join(temp_dir, user_id))
-        os.system('git checkout main')
-        os.system('git pull')
-        # use the above function to copy the files into the git repo for us
-        dst = os.getcwd()
-        if vcs_subdirectory != '':
-            dst = os.path.join(dst, vcs_subdirectory)
-        
-        create_gradeable_submission(src_file, dst)
-        os.system('git add --all')
-        os.system(f"git config user.email '{user_id}@example.com'")
-        os.system(f"git config user.name '{user_id}'")
-        os.system(f"git commit -a --allow-empty -m 'adding submission files' --author='{user_id} <{user_id}@example.com>'")
-        os.system('git push')
-    os.chdir(my_cwd)
-
-def mimic_checkout(repo_path, checkout_path, vcs_subdirectory):
-    os.system(f'git clone {SUBMITTY_DATA_DIR}/vcs/git/{repo_path} {checkout_path}/tmp -b main')
-    if vcs_subdirectory != '':
-        if vcs_subdirectory[0] == '/':
-            vcs_subdirectory = vcs_subdirectory[1:]
-        file_path = os.path.join(f'{checkout_path}/tmp', vcs_subdirectory)
-    else:
-        file_path = os.path.join(f'{checkout_path}/tmp')
-
-    shutil.copytree(file_path, f'{checkout_path}', dirs_exist_ok=True)
-    shutil.rmtree(f'{checkout_path}/tmp')
-
-class User(object):
-    """
-    A basic object to contain the objects loaded from the users.json file. We use this to link
-    against the courses.
-
-    Attributes:
-        id
-        numeric_id
-        password
-        givenname
-        familyname
-        pronouns
-        email
-        group
-        preferred_givenname
-        preferred_familyname
-        access_level
-        registration_section
-        rotating_section
-        unix_groups
-        courses
-    """
-    def __init__(self, user):
-        self.id = user['user_id']
-        self.numeric_id = user['user_numeric_id']
-        self.password = self.id
-        self.givenname = user['user_givenname']
-        self.familyname = user['user_familyname']
-        self.pronouns = user['user_pronouns']
-        self.email = self.id + "@example.com"
-        self.group = 4
-        self.preferred_givenname = None
-        self.preferred_familyname = None
-        self.access_level = 3
-        self.registration_section = None
-        self.rotating_section = None
-        self.grading_registration_section = None
-        self.unix_groups = None
-        self.courses = None
-        self.manual = False
-        self.sudo = False
-
-        if 'user_preferred_givenname' in user:
-            self.preferred_givenname = user['user_preferred_givenname']
-        if 'user_preferred_familyname' in user:
-            self.preferred_familyname = user['user_preferred_familyname']
-        if 'user_email' in user:
-            self.email = user['user_email']
-        if 'user_group' in user:
-            self.group = user['user_group']
-        if self.group < 1 or 4 < self.group:
-            raise SystemExit("ASSERT: user {}, user_group is not between 1 - 4. Check YML file.".format(self.id))
-        if 'user_access_level' in user:
-            self.access_level = user['user_access_level']
-        if self.access_level < 1 or 3 < self.access_level:
-            raise SystemExit("ASSERT: user {}, user_access_level is not between 1 - 3. Check YML file.".format(self.id))
-        if 'registration_section' in user:
-            self.registration_section = int(user['registration_section'])
-        if 'rotating_section' in user:
-            self.rotating_section = int(user['rotating_section'])
-        if 'grading_registration_section' in user:
-            self.grading_registration_section = user['grading_registration_section']
-        if 'unix_groups' in user:
-            self.unix_groups = user['unix_groups']
-        if 'manual_registration' in user:
-            self.manual = user['manual_registration'] is True
-        if 'courses' in user:
-            self.courses = {}
-            if isinstance(user['courses'], list):
-                for course in user['courses']:
-                    self.courses[course] = {"user_group": self.group}
-            elif isinstance(user['courses'], dict):
-                self.courses = user['courses']
-                for course in self.courses:
-                    if 'user_group' not in self.courses[course]:
-                        self.courses[course]['user_group'] = self.group
-            else:
-                raise ValueError("Invalid type for courses key, it should either be list or dict")
-        if 'sudo' in user:
-            self.sudo = user['sudo'] is True
-        if 'user_password' in user:
-            self.password = user['user_password']
-
-    def create(self, force_ssh=False):
-        if not DB_ONLY and not user_exists(self.id):
-            if self.group > 2 and not force_ssh:
-                self.create_non_ssh()
-            else:
-                self.create_ssh()
-            self.create_ldap()
-
-        if self.group <= 1:
-            add_to_group("submitty_course_builders", self.id)
-        if self.sudo:
-            add_to_group("sudo", self.id)
-
-    def create_ssh(self):
-        print("Creating user {}...".format(self.id))
-        os.system("useradd -m -c 'First Last,RoomNumber,WorkPhone,HomePhone' {}".format(self.id))
-        self.set_password()
-
-    def create_non_ssh(self):
-        print("Creating user {}...".format(self.id))
-        os.system("useradd --home /tmp -c \'AUTH ONLY account\' "
-                    "-M --shell /bin/false {}".format(self.id))
-        self.set_password()
-
-    def create_ldap(self):
-        print(f"Creating LDAP user {self.id}...")
-        path = Path("/tmp", self.id)
-        path.write_text(f"""
-dn: uid={self.id},ou=users,dc=vagrant,dc=local
-objectClass: top
-objectClass: account
-objectClass: shadowAccount
-uid: {self.id}
-userPassword: {self.id}
-shadowLastChange: 0
-shadowMax: 0
-shadowWarning: 0""")
-        os.system(f'ldapadd -x -w root_password -D "cn=admin,dc=vagrant,dc=local" -f {path}')
-        path.unlink()
-
-    def set_password(self):
-        print("Setting password for user {}...".format(self.id))
-        os.system("echo {}:{} | chpasswd".format(self.id, self.password))
-
-    def get_detail(self, course, detail):
-        if self.courses is not None and course in self.courses:
-            user_detail = "user_" + detail
-            if user_detail in self.courses[course]:
-                return self.courses[course][user_detail]
-            elif detail in self.courses[course]:
-                return self.courses[course][detail]
-        if detail in self.__dict__:
-            return self.__dict__[detail]
-        else:
-            return None
 
 
 class Course(object):
@@ -2238,20 +1692,6 @@ class Component(object):
             mark.create(self.key, conn, mark_table)
 
 
-class Mark(object):
-    def __init__(self, mark, order):
-        self.note = mark['gcm_note']
-        self.points = mark['gcm_points']
-        self.order = order
-        self.grader = 'instructor'
-        self.key = None
-
-    def create(self, gc_id, conn, table):
-        ins = table.insert().values(gc_id=gc_id, gcm_points=self.points, gcm_note=self.note,
-                                    gcm_order=self.order)
-        res = conn.execute(ins)
-        self.key = res.inserted_primary_key[0]
 
 
-if __name__ == "__main__":
-    main()
+
