+++ .setup/bin/setup_sample_courses.py
@@ -1,0 +1,5 @@
+from sample_courses import main
+
+
+if __name__ == "__main__":
+    main()
